Moved logger class to metrics namespace
