+++ lib/dc/metrics.rb
@@ -1,6 +1,6 @@
 require "dc/metrics/version"
 require "dc/metrics/configuration"
-require "dc/logger"
+require "dc/metrics/logger"
 
 module Dc
   module Metrics
@@ -12,22 +12,22 @@ module Dc
     end
 
     def self.debug(message, metadata)
-      l = Logger.new
+      l = Metrics::Logger.new
       l.log(:debug, message, metadata)
     end
 
     def self.info(message, metadata)
-      l = Logger.new
+      l = Metrics::Logger.new
       l.log(:info, message, metadata)
     end
 
     def self.warn(message, metadata)
-      l = Logger.new
+      l = Metrics::Logger.new
       l.log(:warn, message, metadata)
     end
 
     def self.error(message, metadata)
-      l = Logger.new
+      l = Metrics::Logger.new
       l.log(:error, message, metadata)
     end
